move some traits into enclave-api crate
