+++ modules/enclave-api/src/enclave.rs
@@ -1,11 +1,13 @@
+use crate::errors::Result;
 use crate::EnclaveProtoAPI;
 use host_environment::Environment;
 use lcp_types::Time;
 use sgx_urts::SgxEnclave;
 use std::{marker::PhantomData, ops::DerefMut};
-use store::host::{HostCommitStore, HostStore, HostStoreAccessor};
-use store::transaction::CommitStore;
+use store::host::{HostStore, IntoCommitStore};
+use store::transaction::{CommitStore, CreatedTx, UpdateKey};
 
+/// `Enclave` keeps an enclave id and reference to the host environement
 pub struct Enclave<'e, S: CommitStore> {
     pub(crate) sgx_enclave: SgxEnclave,
     pub(crate) env: &'e Environment,
@@ -29,6 +31,7 @@ where
     }
 }
 
+/// `EnclaveInfo` is an accessor to enclave information
 pub trait EnclaveInfo {
     fn get_home(&self) -> String;
     fn get_eid(&self) -> sgx_types::sgx_enclave_id_t;
@@ -49,10 +52,37 @@ impl<'e, S: CommitStore> EnclaveInfo for Enclave<'e, S> {
     }
 }
 
-impl<'e, S> HostStoreAccessor<S> for Enclave<'e, S>
+/// `HostStoreTxManager` is a transaction manager for the host store
+pub trait HostStoreTxManager<S: CommitStore>: CommitStoreAccessor<S> {
+    /// `begin_tx` creates a transaction and begin it
+    fn begin_tx(&self, update_key: Option<UpdateKey>) -> Result<<S::Tx as CreatedTx>::PreparedTx> {
+        let tx = self.use_mut_store(|store| store.create_transaction(update_key))?;
+        let tx = tx.prepare()?;
+        self.use_mut_store(|store| store.begin(&tx))?;
+        Ok(tx)
+    }
+
+    /// `commit_tx` commits the changes in the transaction
+    fn commit_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) -> Result<()> {
+        self.use_mut_store(|store| store.commit(tx))?;
+        Ok(())
+    }
+
+    /// `rollback_tx` rollbacks the changes in the transaction
+    fn rollback_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) {
+        self.use_mut_store(|store| store.rollback(tx));
+    }
+}
+
+/// `CommitStoreAccessor` is an accessor to the host store
+pub trait CommitStoreAccessor<S: CommitStore> {
+    fn use_mut_store<T>(&self, f: impl FnOnce(&mut S) -> T) -> T;
+}
+
+impl<'e, S> CommitStoreAccessor<S> for Enclave<'e, S>
 where
     S: CommitStore,
-    HostStore: HostCommitStore<S>,
+    HostStore: IntoCommitStore<S>,
 {
     fn use_mut_store<T>(&self, f: impl FnOnce(&mut S) -> T) -> T {
         let mut store = self.env.get_mut_store();
