+++ modules/enclave-api/src/memory.rs
@@ -1,5 +1,7 @@
-use crate::{Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI};
-use store::{host::HostStoreTxManager, memory::MemStore};
+use crate::{
+    enclave::HostStoreTxManager, Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI,
+};
+use store::memory::MemStore;
 
 impl<'e> HostStoreTxManager<MemStore> for Enclave<'e, MemStore> {}
 impl<'e> EnclavePrimitiveAPI<MemStore> for Enclave<'e, MemStore> {}
