+++ modules/enclave-api/src/rocksdb.rs
@@ -1,5 +1,7 @@
-use crate::{Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI};
-use store::{host::HostStoreTxManager, rocksdb::RocksDBStore};
+use crate::{
+    enclave::HostStoreTxManager, Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI,
+};
+use store::rocksdb::RocksDBStore;
 
 impl<'e> HostStoreTxManager<RocksDBStore> for Enclave<'e, RocksDBStore> {}
 impl<'e> EnclavePrimitiveAPI<RocksDBStore> for Enclave<'e, RocksDBStore> {}
