+++ modules/store/src/host.rs
@@ -1,24 +1,23 @@
 use crate::memory::MemStore;
 use crate::prelude::*;
-use crate::transaction::{CommitStore, CreatedTx, TxAccessor, UpdateKey};
+use crate::transaction::{CommitStore, TxAccessor};
 use crate::{KVStore, Result, TxId};
 
+/// `HostStore` defines store implementations on host
 pub enum HostStore {
     #[cfg(feature = "rocksdbstore")]
     RocksDB(crate::rocksdb::RocksDBStore),
     Memory(crate::memory::MemStore),
 }
 
-pub trait HostStoreAccessor<S: CommitStore> {
-    fn use_mut_store<T>(&self, f: impl FnOnce(&mut S) -> T) -> T;
-}
-
-pub trait HostCommitStore<S: CommitStore> {
+/// `IntoCommitStore` converts self into CommitStore
+pub trait IntoCommitStore<S: CommitStore> {
+    /// `apply` applies `f` to CommitStore
     fn apply<T>(&mut self, f: impl FnOnce(&mut S) -> T) -> T;
 }
 
 #[cfg(feature = "rocksdbstore")]
-impl HostCommitStore<crate::rocksdb::RocksDBStore> for HostStore {
+impl IntoCommitStore<crate::rocksdb::RocksDBStore> for HostStore {
     fn apply<T>(&mut self, f: impl FnOnce(&mut crate::rocksdb::RocksDBStore) -> T) -> T {
         match self {
             HostStore::RocksDB(store) => f(store),
@@ -27,7 +26,7 @@ impl HostCommitStore<crate::rocksdb::RocksDBStore> for HostStore {
     }
 }
 
-impl HostCommitStore<MemStore> for HostStore {
+impl IntoCommitStore<MemStore> for HostStore {
     fn apply<T>(&mut self, f: impl FnOnce(&mut MemStore) -> T) -> T {
         match self {
             HostStore::Memory(store) => f(store),
@@ -55,26 +54,5 @@ impl TxAccessor for HostStore {
             HostStore::RocksDB(store) => store.run_in_mut_tx(tx_id, f),
             HostStore::Memory(store) => store.run_in_mut_tx(tx_id, f),
         }
-    }
-}
-
-/// `HostStoreTxManager` is a transaction manager
-pub trait HostStoreTxManager<S: CommitStore>: HostStoreAccessor<S> {
-    /// `begin_tx` begins ...
-    fn begin_tx(&self, update_key: Option<UpdateKey>) -> Result<<S::Tx as CreatedTx>::PreparedTx> {
-        let tx = self.use_mut_store(|store| store.create_transaction(update_key))?;
-        let tx = tx.prepare()?;
-        self.use_mut_store(|store| store.begin(&tx))?;
-        Ok(tx)
-    }
-
-    /// `commit_tx` commits ...
-    fn commit_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) -> Result<()> {
-        self.use_mut_store(|store| store.commit(tx))
-    }
-
-    /// `rollback_tx` rollback ...
-    fn rollback_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) {
-        self.use_mut_store(|store| store.rollback(tx));
     }
 }
